feat(dotfiles): collect.py now deletes only if needed, new configs added

+++ collect.py
@@ -12,18 +12,17 @@ def copy_to_config_directory(source_directory):
     
     destination_path = os.path.join(config_directory, item_name)
     
-    if os.path.exists(destination_path):
-        if os.path.isdir(destination_path):
-            shutil.rmtree(destination_path)
-        else:
-            os.remove(destination_path)
     try:
         if os.path.isdir(source_directory):
+            if os.path.exists(destination_path):
+                if os.path.isdir(destination_path):
+                    shutil.rmtree(destination_path)
+                else:
+                    os.remove(destination_path)
             shutil.copytree(source_directory, destination_path)
             print(f"Copied directory {source_directory} to {destination_path}")
         else:
-            shutil.copy2(source_directory, destination_path)
-            print(f"Copied file {source_directory} to {destination_path}")
+            print(f"Not found: {source_directory}")
     except Exception as e:
         print(f"An error occurred: {e}")
 
@@ -36,22 +35,25 @@ def copy_rc_files(rc_files):
     for rc_file in rc_files:
         source_file = os.path.join(home_directory, rc_file)
         destination_file = os.path.join(config_directory, rc_file)
-        if os.path.exists(destination_file):
-            os.remove(destination_file)
         try:
+            if os.path.exists(source_file):
+                if os.path.exists(destination_file):
+                    os.remove(destination_file)
             shutil.copy2(source_file, destination_file)
             print(f"Copied {rc_file} to {destination_file}")
         except FileNotFoundError:
-            print(f"{rc_file} not found in the home directory.")
+            print(f"Not found: {rc_file}")
         except Exception as e:
-            print(f"An error occurred while copying {rc_file}: {e}")
+            print(f"An error occurred: {e}")
 
 configs_to_copy = [
         'alacritty',
         'dwm',
         'nvim',
         'ranger',
-        'tmux'
+        'tmux',
+        'rofi',
+        'waybar'
         ]
 for source_path in configs_to_copy:
     source_path = '/home/w/.config/' + source_path
